+++ Packages/com.mimylab.fukuroudon/Runtime/AmbientEffectAssistant/Scripts/AreaCulling.cs
@@ -28,7 +28,6 @@ namespace MimyLab.FukuroUdon
         [SerializeField, Tooltip("Only Sphere, Capsule, Box, and Convexed Mesh Colliders")]
         private Collider[] _area = new Collider[0];
 
-        private Vector3 _viewPointPosition;
         private bool _prevEnabled;
 
         private bool _initialized = false;
@@ -46,13 +45,10 @@ namespace MimyLab.FukuroUdon
         {
             Initialize();
 
-            if (position == _viewPointPosition) { return; }
-
-            SnapViewPoint(position);
-            _viewPointPosition = position;
+            CullingByViewPointPosition(position);
         }
 
-        private void SnapViewPoint(Vector3 vpPosition)
+        private void CullingByViewPointPosition(Vector3 vpPosition)
         {
             var isIn = false;
             //var nearest = Vector3.positiveInfinity;