+++ Packages/com.mimylab.fukuroudon/Runtime/AmbientEffectAssistant/Scripts/FlexibleReverbZone.cs
@@ -23,8 +23,6 @@ namespace MimyLab.FukuroUdon
 
         Transform _reverbZoneTransform;
 
-        private Vector3 _viewPointPosition;
-
         private bool _initialized = false;
         private void Initialize()
         {
@@ -39,10 +37,7 @@ namespace MimyLab.FukuroUdon
         {
             Initialize();
 
-            if (position == _viewPointPosition) { return; }
-
             SnapViewPointPosition(position);
-            _viewPointPosition = position;
         }
 
         private void SnapViewPointPosition(Vector3 vpPosition)