+++ Packages/com.mimylab.fukuroudon/Runtime/AmbientEffectAssistant/Scripts/FlexibleSpatialAudio.cs
@@ -29,8 +29,6 @@ namespace MimyLab.FukuroUdon
         private Transform _decayTransform;
         private Transform _innerTransform;
 
-        private Vector3 _viewPointPosition;
-
         private bool _initialized = false;
         private void Initialize()
         {
@@ -46,10 +44,7 @@ namespace MimyLab.FukuroUdon
         {
             Initialize();
 
-            if (position == _viewPointPosition) { return; }
-
             SnapViewPointPosition(position);
-            _viewPointPosition = position;
         }
 
         private void SnapViewPointPosition(Vector3 vpPosition)