+++ Packages/com.mimylab.fukuroudon/Runtime/AmbientEffectAssistant/Scripts/ViewPointTracker.cs
@@ -29,21 +29,31 @@ namespace MimyLab.FukuroUdon
         [SerializeField]
         private UdonBehaviour[] _rotationReceiver = new UdonBehaviour[0];
 
-        private VRCPlayerApi _lPlayer;
+        private VRCPlayerApi _localPlayer;
         private Vector3 _prevViewPointPosition;
         private Quaternion _prevViewPointRotation;
 
         private void Start()
         {
-            _lPlayer = Networking.LocalPlayer;
-            var viewPoint = _lPlayer.GetTrackingData(VRCPlayerApi.TrackingDataType.Head);
+            _localPlayer = Networking.LocalPlayer;
+            var viewPoint = _localPlayer.GetTrackingData(VRCPlayerApi.TrackingDataType.Head);
             _prevViewPointPosition = viewPoint.position;
             _prevViewPointRotation = viewPoint.rotation;
+            this.transform.SetPositionAndRotation(_prevViewPointPosition, _prevViewPointRotation);
+
+            foreach (var target in _viewPointReceiver)
+            {
+                if (target)
+                {
+                    target.viewPointTracker = this.transform;
+                    target.OnViewPointChanged();
+                }
+            }
         }
 
         public override void PostLateUpdate()
         {
-            var viewPoint = _lPlayer.GetTrackingData(VRCPlayerApi.TrackingDataType.Head);
+            var viewPoint = _localPlayer.GetTrackingData(VRCPlayerApi.TrackingDataType.Head);
             var viewPointPosition = viewPoint.position;
             var viewPointRotation = viewPoint.rotation;
 
@@ -52,9 +62,11 @@ namespace MimyLab.FukuroUdon
 
             if (isMoved | isTurned)
             {
+                this.transform.SetPositionAndRotation(viewPointPosition, viewPointRotation);
+
                 foreach (var target in _viewPointReceiver)
                 {
-                    if (target) target.ReceiveViewPoint(viewPointPosition, viewPointRotation);
+                    if (target) target.OnViewPointChanged();
                 }
             }