+++ Packages/com.mimylab.fukuroudon/CHANGELOG.md
@@ -6,6 +6,12 @@
 
 利用手順は[こちら](https://github.com/mimyquality/FukuroUdon/wiki)からご確認ください。
 
+## [2.0.0-beta.2] - 2024/9/14
+
+- **Changed**
+  - AmbientEffect Assistant
+    - 軽量化。これに伴い、ViewPoint Tracker自身がビューポイントと一致して動くようになりました。
+
 ## [2.0.0-beta] - 2024/8/30
 
 - **Changed**
@@ -516,6 +522,7 @@
     - 0.2.7バージョンで、later-joiner視点でjoin直後が非アクティブだと同期しなかったのを修正しました。
     - 他人がピックアップした際にオブジェクトが一瞬跳ねる現象を低減しました。
 
+[2.0.0-beta.2]: https://github.com/mimyquality/FukuroUdon/releases/tag/2.0.0-beta.2
 [2.0.0-beta]: https://github.com/mimyquality/FukuroUdon/releases/tag/2.0.0-beta
 [1.24.0]: https://github.com/mimyquality/FukuroUdon/releases/tag/1.24.0
 [1.23.0]: https://github.com/mimyquality/FukuroUdon/releases/tag/1.23.0